PoC: Improve TPS for voter app
This patch improves the throughput of the voter app from 7-8k TPS on a
single core to 11-12 TPS on my machine. It does it by combining multiple
outgoing messages into a single buffer. Previously each message (~50
bytes) was sent individually, so a lot of time was wasted in libuv and
kernel space to deal with the overhead of sending many messages this
small.

NOTE: This patch needs more work. You probably want to add a timer that
flushes pending messages after a while if the configured threshold is
not reached. The patch also error handling in case socket.write() fails
(there needs to be a try..catch block in the _flush method).